Break apart CommonTools. Its methods put it correct peer classes, or into new service classes.
