+++ src/server/src/test/java/io/cassandrareaper/resources/ClusterResourceTest.java
@@ -26,6 +26,7 @@ import io.cassandrareaper.storage.MemoryStorage;
 import java.net.URI;
 import java.time.Duration;
 import java.util.Arrays;
+import java.util.Set;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
@@ -163,6 +164,15 @@ public final class ClusterResourceTest {
         mocks.context.storage.getRepairSchedulesForClusterAndKeyspace(
             CLUSTER_NAME, "keyspace1"))
         .hasSize(1);
+  }
+
+  @Test
+  public void testParseSeedHost() {
+    String seedHostStringList = "127.0.0.1 , 127.0.0.2,  127.0.0.3";
+    Set<String> seedHostSet = ClusterResource.parseSeedHosts(seedHostStringList);
+    Set<String> seedHostExpectedSet = Sets.newHashSet("127.0.0.2", "127.0.0.1", "127.0.0.3");
+
+    assertEquals(seedHostSet, seedHostExpectedSet);
   }
 
   private MockObjects initMocks() throws ReaperException {
