+++ src/server/src/test/java/io/cassandrareaper/service/SegmentRunnerTest.java
@@ -419,4 +419,76 @@ public final class SegmentRunnerTest {
     assertFalse(SegmentRunner.okToRepairSegment(false, false, DatacenterAvailability.EACH));
     assertTrue(SegmentRunner.okToRepairSegment(true, true, DatacenterAvailability.EACH));
   }
+
+  @Test
+  public void getTablesToRepairRemoveOneTableTest() throws ReaperException {
+    JmxProxy coord = mock(JmxProxy.class);
+    when(coord.getTableNamesForKeyspace(Mockito.anyString()))
+        .thenReturn(Sets.newHashSet("table1", "table2", "table3"));
+
+
+    RepairUnit unit = mock(RepairUnit.class);
+    when(unit.getBlacklistedTables()).thenReturn(Sets.newHashSet("table1"));
+    when(unit.getColumnFamilies()).thenReturn(Sets.newHashSet());
+    when(unit.getKeyspaceName()).thenReturn("test");
+
+    assertEquals(Sets.newHashSet("table2", "table3"), SegmentRunner.getTablesToRepair(coord, unit));
+  }
+
+  @Test
+  public void getTablesToRepairRemoveTwoTablesTest() throws ReaperException {
+    JmxProxy coord = mock(JmxProxy.class);
+    when(coord.getTableNamesForKeyspace(Mockito.anyString()))
+        .thenReturn(Sets.newHashSet("table1", "table2", "table3"));
+
+    RepairUnit unit = mock(RepairUnit.class);
+    when(unit.getBlacklistedTables()).thenReturn(Sets.newHashSet("table1", "table3"));
+    when(unit.getColumnFamilies()).thenReturn(Sets.newHashSet());
+    when(unit.getKeyspaceName()).thenReturn("test");
+
+    assertEquals(Sets.newHashSet("table2"), SegmentRunner.getTablesToRepair(coord, unit));
+  }
+
+  @Test
+  public void getTablesToRepairRemoveOneTableFromListTest() throws ReaperException {
+    JmxProxy coord = mock(JmxProxy.class);
+    when(coord.getTableNamesForKeyspace(Mockito.anyString()))
+        .thenReturn(Sets.newHashSet("table1", "table2", "table3"));
+
+    RepairUnit unit = mock(RepairUnit.class);
+    when(unit.getBlacklistedTables()).thenReturn(Sets.newHashSet("table1"));
+    when(unit.getColumnFamilies()).thenReturn(Sets.newHashSet("table1", "table2"));
+    when(unit.getKeyspaceName()).thenReturn("test");
+
+    assertEquals(Sets.newHashSet("table2"), SegmentRunner.getTablesToRepair(coord, unit));
+  }
+
+  @Test(expected = IllegalStateException.class)
+  public void getTablesToRepairRemoveAllFailingTest() throws ReaperException {
+    JmxProxy coord = mock(JmxProxy.class);
+    when(coord.getTableNamesForKeyspace(Mockito.anyString()))
+        .thenReturn(Sets.newHashSet("table1", "table2", "table3"));
+
+    RepairUnit unit = mock(RepairUnit.class);
+    when(unit.getBlacklistedTables()).thenReturn(Sets.newHashSet("table1", "table2", "table3"));
+    when(unit.getColumnFamilies()).thenReturn(Sets.newHashSet());
+    when(unit.getKeyspaceName()).thenReturn("test");
+
+    SegmentRunner.getTablesToRepair(coord, unit);
+  }
+
+  @Test(expected = IllegalStateException.class)
+  public void getTablesToRepairRemoveAllFromListFailingTest() throws ReaperException {
+    JmxProxy coord = mock(JmxProxy.class);
+    when(coord.getTableNamesForKeyspace(Mockito.anyString()))
+        .thenReturn(Sets.newHashSet("table1", "table2", "table3", "table4"));
+
+    RepairUnit unit = mock(RepairUnit.class);
+    when(unit.getBlacklistedTables()).thenReturn(Sets.newHashSet("table1", "table2", "table3"));
+    when(unit.getColumnFamilies()).thenReturn(Sets.newHashSet("table1", "table2", "table3"));
+    when(unit.getKeyspaceName()).thenReturn("test");
+
+    SegmentRunner.getTablesToRepair(coord, unit);
+  }
+
 }
